+++ R/reformulare_ATSP_as_TSP.R
@@ -36,14 +36,26 @@
 #' distance coming from the city. The distances between all cities and the
 #' distances between all dummy cities are set to `infeasible`, a very
 #' large value which prevents the solver from using these links.
+#' We use infinite values here and [solve_TSP()] treats them appropriately.
 #'
+#' `filter_ATSP_as_TSP_dummies()` can be used to extract the solution for the original
+#' ATSP from the tour found for an ATSP reformulated as a TSP. Note that the symmetric TSP
+#' tour does not reveal the direction for the ATSP. The filter function computed the
+#' tour length for both directions and returns the shorter tour.
+#'
+#' [solve_TSP()] has a parameter `as_TSP` which preforms the reformulation and
+#' filtering the dummy cities automatically.
+#'
+#' **Note on performance:** Doubling the problem size is a performance issue especially
+#' has a negative impact on solution quality for heuristics. It should only be used
+#' together with Concorde when the optimal solution is required. Most heuristics can solve
+#' ATSPs directly with good solution quality.
 #' @family TSP
 #'
 #' @param x an [ATSP].
 #' @param infeasible value for infeasible connections.
 #' @param cheap value for distance between a city and its corresponding dummy
-#' city. This needs to be very small compared to the smallest distance. Value
-#' `"auto"` calculates cheap as mimimum distance minus twice the distance range.
+#' city.
 #' @param tour a [TOUR] created for a ATSP reformulated as a TSP.
 #' @param atsp the original [ATSP].
 #' @return
@@ -61,41 +73,40 @@
 #' austin <- which(labels(USCA50) == "Austin, TX")
 #' atsp <- as.ATSP(USCA50)
 #' atsp[, austin] <- 0
+#' atsp
 #'
-#' ## reformulate as a TSP (by adding dummy cities marked with *)
+#' ## reformulate as a TSP (by doubling the number of cities with dummy cities marked with *)
 #' tsp <- reformulate_ATSP_as_TSP(atsp)
-#' labels(tsp)
+#' tsp
 #'
-#' ## create tour for the TSP (now you can use Concorde to find the optimal solution)
+#' ## create tour for the TSP. You should use Concorde to find the optimal solution.
 #' # tour_tsp <- solve_TSP(tsp, method = "concorde")
-#' ## The standard heuristic is not very good for this problem.
+#' # The standard heuristic is bad for this problem. We use it here because
+#' #   Concord may not be installed.
 #' tour_tsp <- solve_TSP(tsp)
 #' head(labels(tour_tsp), n = 10)
 #' tour_tsp
-#' ## Note that the tour length is not valid since it includes the cheap segments with
-#' ## large negative costs.
+#' # The tour length is -Inf since it includes cheap links
+#' #  from a city to its dummy city.
 #'
-#' ## get the solution for the original ATSP by filtering out the dummy cities
+#' ## get the solution for the original ATSP by filtering out the dummy cities.
 #' tour_atsp <- filter_ATSP_as_TSP_dummies(tour_tsp, atsp = atsp)
 #' tour_atsp
 #' head(labels(tour_atsp), n = 10)
 #'
-#' ## This process can also be done using as_TSP = TRUE:
-#' # solve_TSP(atsp, as_TSP = TRUE, control = list(rep = 10))
+#' ## This process can also be done automatically by using as_TSP = TRUE:
+#' # solve_TSP(atsp, method = "concorde", as_TSP = TRUE)
 #'
-#' ## The default heuristic can directly solve ATSPs with good results
-#' ## (much better than on the reformulated problem).
+#' ## The default heuristic can directly solve ATSPs with results close to the
+#' #  optimal solution of 12715.
 #' solve_TSP(atsp, control = list(rep = 10))
 #' @export
 reformulate_ATSP_as_TSP <-
   function(x,
     infeasible = Inf,
-    cheap = "auto") {
+    cheap = -Inf) {
     if (!inherits(x, "ATSP"))
       stop("x is not an ATSP object!")
-
-    if (cheap == "auto")
-      cheap <- min(x) - abs(diff(range(x)))*2
 
     method <- attr(x, "method")
     m <- as.matrix(x)
@@ -120,7 +131,6 @@ reformulate_ATSP_as_TSP <-
     lab <- c(labels(x), paste(labels(x), "*", sep = ""))
     dimnames(tsp) <- list(lab, lab)
     attr(tsp, "method") <- method
-
 
     ## return as TSP
     TSP(tsp)