+++ NEWS.md
@@ -1,9 +1,11 @@
 # TSP 1.2-0-1 (xx/xx/2022)
 
 ## Changes
-* added filter_ATSP_as_TSP_dummies()
+* added filter_ATSP_as_TSP_dummies().
 * tour_length() now also accepts integer vectors instead of TOUR.
 
+## Bugfix
+* Integer values are now left unscaled for Concorde.
 
 # TSP 1.2-0 (02/21/2022)